requirements.txt
python-dotenv
openai
requests
websocket-client
